// src/app/core/enum/law-enum.ts
export enum LawEnum {
  SOCIETARIO = 'SOCIETARIO',
  LAVORO = 'LAVORO',
  COMMERCIALE = 'COMMERCIALE',
  CIVILE = 'CIVILE',
  IMMOBILIARE = 'IMMOBILIARE',
  CONTENZIOSO = 'CONTENZIOSO',
  TRIBUTARIO = 'TRIBUTARIO',
  PENALE = 'PENALE',
  MEDIAZIONE = 'MEDIAZIONE',
  PROGETTAZIONE_BANDI_NAZIONALI_EUROPEI = 'PROGETTAZIONE_BANDI_NAZIONALI_EUROPEI',
  CONSULENZA_LEGALE_CONTINUATIVA = 'CONSULEZA_LEGALE_CONTINUATIVA',
}